refactor(adr-061): classify 20 dumper_*.py modules into extract layer
Classify 20 of the 23 dumper*.py header-AST/DWARF parsing-engine modules
into the `extract` responsibility layer (architecture/modules.yaml):
dumper_ast_config.py, dumper_ast_config_cpp20.py,
dumper_ast_config_cpp20_chains.py, dumper_cache.py,
dumper_castxml_probe.py, dumper_castxml_typedefs.py,
dumper_clang_attributes.py, dumper_clang_errors.py,
dumper_clang_qualifiers.py, dumper_clang_streaming.py,
dumper_clang_vtable.py, dumper_contract.py, dumper_debug.py,
dumper_elf_fallback.py, dumper_elf_symbols.py, dumper_layout_backfill.py,
dumper_manifest.py, dumper_sysinc.py, dumper_toolchain.py
(dumper_clang.py/dumper_scoping.py were already classified).

Route the resulting frontends -> extract violations through the existing
abicheck/workflows/extraction.py facade (mirrors PR #907's pattern):
cli_dump_helpers.py, cli_dump_non_elf.py, and compat/cli.py now import
ast_memoize_scope, suppress_streaming_prune, and
_manifest_declared_includes through the facade instead of directly.

Update one CLI_CONTRACT_ALLOWLIST line-pin in check_ai_readiness.py for a
shifted line number (import edits moved dumper.dump's call site by 3
lines); no allowlist entries added or removed.

Deliberately left unclassified, with reasoning recorded in the changelog
fragment:
- dumper.py: probe_harness.py (classified `compare`) imports `dump` at
  function scope, and compare's may_import is [model] only -- no facade
  can bridge that edge without reclassifying probe_harness.py itself.
- dumper_hybrid.py: imports diff_cxx_rules/diff_helpers (compare) directly
  -- extract's may_import is [model, storage], compare is unreachable.
- dumper_castxml.py: three compare-classified diff_*.py modules import its
  synthetic ctor/dtor-key constants directly -- same structural block.

Verified: scripts/check_architecture.py (0 errors), scripts/check_ai_readiness.py
(0 errors, 146 warnings, unchanged from baseline), mypy abicheck/ (clean),
scripts/adr_status_sync.py (clean), full fast unit suite (32966 passed, 0
failed, 130 skipped, 4 xfailed).
