+++ architecture/modules.yaml
@@ -85,8 +85,27 @@
         "abicheck/buildsource/type_graph.py",
         "abicheck/debug_resolver.py",
         "abicheck/dump_manifest.py",
+        "abicheck/dumper_ast_config.py",
+        "abicheck/dumper_ast_config_cpp20.py",
+        "abicheck/dumper_ast_config_cpp20_chains.py",
+        "abicheck/dumper_cache.py",
+        "abicheck/dumper_castxml_probe.py",
+        "abicheck/dumper_castxml_typedefs.py",
         "abicheck/dumper_clang.py",
+        "abicheck/dumper_clang_attributes.py",
+        "abicheck/dumper_clang_errors.py",
+        "abicheck/dumper_clang_qualifiers.py",
+        "abicheck/dumper_clang_streaming.py",
+        "abicheck/dumper_clang_vtable.py",
+        "abicheck/dumper_contract.py",
+        "abicheck/dumper_debug.py",
+        "abicheck/dumper_elf_fallback.py",
+        "abicheck/dumper_elf_symbols.py",
+        "abicheck/dumper_layout_backfill.py",
+        "abicheck/dumper_manifest.py",
         "abicheck/dumper_scoping.py",
+        "abicheck/dumper_sysinc.py",
+        "abicheck/dumper_toolchain.py",
         "abicheck/dwarf_advanced.py",
         "abicheck/dwarf_metadata.py",
         "abicheck/elf_metadata.py",
